+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 * Un hide permission 'Invoice: Cancel invoices'. Refs UINV-362.
+* Previous invoice is fetched when Approve and Pay action is called. Refs UINV-372.
 
 ## [3.1.0](https://github.com/folio-org/ui-invoice/tree/v3.1.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v3.0.2...v3.1.0)